Refactor CreationTools mouseReleased method to be more readable
